requirements.txt
aiofiles==0.5.0
alembic==1.4.3
aniso8601==8.0.0
appdirs==1.4.4
attrs==20.2.0
black==20.8b1
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
gunicorn==20.0.4
h11==0.11.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.1
hyperframe==6.0.0
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kiwisolver==1.2.0
libcst==0.3.13
lxml==4.5.0
Mako==1.1.3
manage.py==0.2.10
MarkupSafe==1.1.1
marshmallow==3.8.0
matplotlib==3.3.2
MonkeyType==20.5.0
multidict==5.0.0
mypy-extensions==0.4.3
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pathspec==0.8.0
Pillow==8.0.1
pkg-resources==0.0.0
pluggy==0.13.1
priority==1.3.0
psycopg2-binary==2.8.6
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.1
pytest-flask-sqlalchemy==1.0.2
pytest-mock==3.3.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
Quart==0.13.1
quart-openapi==1.6.0
regex==2020.10.23
requests==2.24.0
requests-futures==1.0.0
selenium==3.141.0
six==1.15.0
sortedcontainers==2.2.2
SQLAlchemy==1.3.20
toml==0.10.1
tqdm==4.51.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
Werkzeug==1.0.1
wsproto==0.15.0
yahooquery==2.2.8
zipp==3.4.0
zmq==0.0.0
